Split the connection out into separate files
